Arrumando pagina /home, simulados e protegendo rotas

+++ routes/web.php
@@ -12,23 +12,26 @@
 */
 
 Route::get('/', function () {
-    return view('welcome');
+    return redirect('/login');
 });
 
 Auth::routes();
 
-Route::get('/home', 'HomeController@index')->name('home');
+Route::get('/home', 'UsuarioController@index')->middleware('auth')->name('home');
+Route::get('/cursos/{id}', 'UsuarioController@curso')->middleware('auth');
+Route::get('/arquivos/{tipo}/{id}', 'ArquivoController@show')->middleware('auth');
 
-Route::group(['middleware' => ['auth', 'admin'], 'prefix' => 'admin'], function () {
+Route::group(['middleware' => ['auth', 'admin'], 'prefix' => 'admin'],function () {
 	Route::get('/home', 'AdminController@index')->name('painel-admin');
-	Route::get('/cursos/{id}', 'AdminController@curso')->name('cursos-admin');
+	Route::get('/cursos/{id}', 'AdminController@curso')->name('curso-admin');
 });
 
-Route::get('/cursos', 'CursoController@index');
-Route::get('/cursos/{id}', 'CursoController@show');
-Route::get('/cursos/{id}/simulados/gabaritos', 'CursoController@showSimuladosGabarito');
 
-Route::post('/simulados', 'SimuladoController@store');
-Route::delete('/simulados/{id}', 'SimuladoController@destroy');
+Route::group(['middleware' => 'auth', 'prefix' => 'api'],function () {
+	Route::get('/cursos', 'CursoController@index');
+	Route::get('/cursos/{id}/simulados/gabaritos', 'CursoController@showSimuladosGabarito');
 
-Route::get('/arquivos/{tipo}/{id}', 'ArquivoController@show');
+	Route::post('/simulados', 'SimuladoController@store')->middleware('admin');
+	Route::delete('/simulados/{id}', 'SimuladoController@destroy')->middleware('admin');
+
+});
